change to execute

+++ scripts/utils.sh
@@ -93,14 +93,6 @@ function DEBUG_END() {
     DEBUG set +x
 }
 
-function DRYRUN() {
-    if [ "$_DRYRUN" == "on" ]; then
-        echo "DRYRUN: $@"
-    else
-        $@
-    fi
-}
-
 function execute() {
     local -r CMDS="$1"
     local -r MSG="${2:-$1}"
@@ -109,51 +101,38 @@ function execute() {
     local exitCode=0
     local cmdsPID=""
 
-    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+    if [ "$_DRYRUN" == "on" ]; then
+        echo "DRYRUN: $CMDS"
+    else
+        # If the current process is ended,
+        # also end all its subprocesses.
+        set_trap "EXIT" "kill_all_subprocesses"
 
-    # If the current process is ended,
-    # also end all its subprocesses.
-
-    set_trap "EXIT" "kill_all_subprocesses"
-
-    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
-    # Execute commands in background
-
-    eval "$CMDS" \
+        # Execute commands in background
+        eval "$CMDS" \
         &> /dev/null \
         2> "$TMP_FILE" &
 
-    cmdsPID=$!
+        cmdsPID=$!
 
-    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+        # Show a spinner if the commands
+        # require more time to complete.
+        show_spinner "$cmdsPID" "$CMDS" "$MSG"
 
-    # Show a spinner if the commands
-    # require more time to complete.
+        # Wait for the commands to no longer be executing
+        # in the background, and then get their exit code.
+        wait "$cmdsPID" &> /dev/null
+        exitCode=$?
 
-    show_spinner "$cmdsPID" "$CMDS" "$MSG"
+        # Print output based on what happened.
+        print_result $exitCode "$MSG"
 
-    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+        if [ $exitCode -ne 0 ]; then
+            print_error_stream < "$TMP_FILE"
+        fi
 
-    # Wait for the commands to no longer be executing
-    # in the background, and then get their exit code.
-
-    wait "$cmdsPID" &> /dev/null
-    exitCode=$?
-
-    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
-    # Print output based on what happened.
-
-    print_result $exitCode "$MSG"
-
-    if [ $exitCode -ne 0 ]; then
-        print_error_stream < "$TMP_FILE"
+        rm -rf "$TMP_FILE"
     fi
-
-    rm -rf "$TMP_FILE"
-
-    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
     return $exitCode
 }
@@ -180,8 +159,6 @@ function get_os() {
 function get_os_version() {
     local os=""
     local version=""
-
-    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
     os="$(get_os)"
